Vault bugfix by IsS127 & Multi Module Project
