+++ nameless-bungee/src/main/java/com/namelessmc/namelessplugin/bungeecord/NamelessPlugin.java
@@ -11,6 +11,7 @@ import com.namelessmc.namelessplugin.bungeecord.commands.ReportCommand;
 import com.namelessmc.namelessplugin.bungeecord.mcstats.Metrics;
 import com.namelessmc.namelessplugin.bungeecord.player.PlayerEventListener;
 
+import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Plugin;
 import net.md_5.bungee.config.Configuration;
@@ -18,23 +19,10 @@ import net.md_5.bungee.config.ConfigurationProvider;
 import net.md_5.bungee.config.YamlConfiguration;
 
     /*
-     * Bungeecord version made by IsS127
+     *  Bungeecord Version by IsS127
      */
 
 public class NamelessPlugin extends Plugin {
-	/*
-	 *  Colors you can use to the console :)
-	 */
-	public final String COLOR_BLACK = "\u001B[30m";
-	public final String COLOR_RED = "\u001B[31m";
-	public final String COLOR_GREEN = "\u001B[32m";
-	public final String COLOR_CYAN = "\u001B[36m";
-	public final String COLOR_YELLOW = "\u001B[33m";
-	public final String COLOR_BLUE = "\u001B[34m";
-	public final String COLOR_PURPLE = "\u001B[35m";
-	public final String COLOR_WHITE = "\u001B[37m";
-	// Must use this after writing a line. 
-	public final String COLOR_RESET = "\u001B[0m";
 	
 	/*
 	 *  API URL
@@ -80,7 +68,7 @@ public class NamelessPlugin extends Plugin {
 		try {
             metrics = new Metrics(this);
             metrics.start();
-            getLogger().info(COLOR_CYAN + "Metrics Started!" + COLOR_RESET);
+            getLogger().info(ChatColor.translateAlternateColorCodes('&', "&3Metrics Started!&r"));
         } catch (IOException e) {
             e.printStackTrace();
         } 
@@ -123,11 +111,11 @@ public class NamelessPlugin extends Plugin {
 			if(!file.exists()){
 				try (InputStream in = getResourceAsStream("config.yml")) {
 					// Config doesn't exist, create one now...
-					getLogger().info(COLOR_BLUE + "Creating NamelessMC configuration file..." + COLOR_RESET);
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&1Creating NamelessMC configuration file...&r"));
                     Files.copy(in, file.toPath());
 					
-					getLogger().info(COLOR_RED + "NamelessMC needs configuring, disabling features..." + COLOR_RESET);
-					getLogger().info(COLOR_RED + "Please Configure NamelessMC config.yml!" + COLOR_RESET);
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4NamelessMC needs configuring, disabling features...&r"));
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4Please Configure NamelessMC config.yml!&r"));
 					hasSetUrl = false;
 					
                 } catch (IOException e) {
@@ -139,13 +127,13 @@ public class NamelessPlugin extends Plugin {
 				
 				
 				// Exists already, load it
-				getLogger().info(COLOR_GREEN + "Loading NamelessMC configuration file..." + COLOR_RESET);
+				getLogger().info(ChatColor.translateAlternateColorCodes('&', "&2Loading NamelessMC configuration file...&r"));
 				
 				apiURL = config.getString("api-url");
 				
 				if(apiURL.isEmpty()){
 					// API URL not set
-					getLogger().info(COLOR_RED + "No API URL set in the NamelessMC configuration, disabling features." + COLOR_RESET);
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4No API URL set in the NamelessMC configuration, disabling features.&r"));
 					hasSetUrl = false;
 				}
 			}
