+++ nameless-spigot/src/main/java/com/namelessmc/namelessplugin/spigot/NamelessPlugin.java
@@ -4,6 +4,7 @@ package com.namelessmc.namelessplugin.spigot;
 import java.io.File;
 import java.io.IOException;
 
+import org.bukkit.ChatColor;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.RegisteredServiceProvider;
@@ -18,19 +19,6 @@ import com.namelessmc.namelessplugin.spigot.player.PlayerEventListener;
 import net.milkbowl.vault.permission.Permission;
 
 public class NamelessPlugin extends JavaPlugin {
-	/*
-	 *  Colors you can use to the console :)
-	 */
-	public final String COLOR_BLACK = "\u001B[30m";
-	public final String COLOR_RED = "\u001B[31m";
-	public final String COLOR_GREEN = "\u001B[32m";
-	public final String COLOR_CYAN = "\u001B[36m";
-	public final String COLOR_YELLOW = "\u001B[33m";
-	public final String COLOR_BLUE = "\u001B[34m";
-	public final String COLOR_PURPLE = "\u001B[35m";
-	public final String COLOR_WHITE = "\u001B[37m";
-	// Must use this after writing a line. 
-	public final String COLOR_RESET = "\u001B[0m";
 	
 	/*
 	 * Metrics
@@ -44,14 +32,20 @@ public class NamelessPlugin extends JavaPlugin {
 	private String apiURL = "";
 	
 	/*
-	 *  Is Vault integration enabled?
+	 *  Vault Integration
 	 */
 	private boolean useVault = false;
 	
 	/*
-	 *  Vault permissions
+	 *  Vault Permissions
 	 */
 	private Permission permissions = null;
+	
+	/*
+	 *  Groups Support 
+	 */
+	@SuppressWarnings("unused")
+	private boolean useGroups = false;
 	
 	/*
 	 *  Enable reports?
@@ -64,7 +58,7 @@ public class NamelessPlugin extends JavaPlugin {
 	private boolean isDisabled = false;
 	
 	/*
-	 *  NameLessMC permission string.
+	 *  NamelessMC permissions strings.
 	 */
 	
 	public final String permission = "namelessmc";
@@ -101,7 +95,7 @@ public class NamelessPlugin extends JavaPlugin {
 		try {
             metrics = new Metrics(this);
             metrics.start();
-            getLogger().info(COLOR_CYAN + "Metrics Started!" + COLOR_RESET);
+            getLogger().info(ChatColor.translateAlternateColorCodes('&', "&3Metrics Started!&r"));
         } catch (IOException e) {
             e.printStackTrace();
         } 
@@ -123,11 +117,18 @@ public class NamelessPlugin extends JavaPlugin {
 	 */
 	public void detectVault(){
 				if(getServer().getPluginManager().getPlugin("Vault") != null){
-					// Installed
+					// Enable Vault integration and setup Permissions.
 					useVault = true;
 					initPermissions();
+					// Check if the permissions plugin has groups.
+					if(permissions.hasGroupSupport()){
+						useGroups = true;
+					} else {
+						getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4Permissions plugin does NOT support groups! Disabling NamelessMC group synchronisation.&r"));
+						useGroups = false;
+					}
 				} else {
-					getLogger().info(COLOR_RED + "Couldn't detect Vault, disabling NamelessMC group synchronisation." + COLOR_RESET);
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4Couldn't detect Vault, disabling NamelessMC Vault integration.&r"));
 				}
 	}
 	
@@ -146,10 +147,10 @@ public class NamelessPlugin extends JavaPlugin {
 			
 			if(!file.exists()){
 				// Config doesn't exist, create one now...
-				getLogger().info(COLOR_BLUE + "Creating NamelessMC configuration file..." + COLOR_RESET);
+				getLogger().info(ChatColor.translateAlternateColorCodes('&', "&1Creating NamelessMC configuration file...&r"));
 				this.saveDefaultConfig();
 				
-				getLogger().info(COLOR_RED + "NamelessMC needs configuring, disabling..." + COLOR_RESET);
+				getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4NamelessMC needs configuring, disabling...&r"));
 				
 				// Disable plugin
 				getServer().getPluginManager().disablePlugin(this);
@@ -164,12 +165,12 @@ public class NamelessPlugin extends JavaPlugin {
 				
 				
 				// Exists already, load it
-				getLogger().info(COLOR_GREEN + "Loading NamelessMC configuration file..." + COLOR_RESET);
+				getLogger().info(ChatColor.translateAlternateColorCodes('&', "&2Loading NamelessMC configuration file...&r"));
 				
 				apiURL = yamlConfigFile.getString("api-url");
 				if(apiURL.isEmpty()){
 					// API URL not set
-					getLogger().info(COLOR_RED + "No API URL set in the NamelessMC configuration, disabling..." + COLOR_RESET);
+					getLogger().info(ChatColor.translateAlternateColorCodes('&', "&4No API URL set in the NamelessMC configuration, disabling...&r"));
 					getServer().getPluginManager().disablePlugin(this);
 				}
 				
@@ -211,6 +212,7 @@ public class NamelessPlugin extends JavaPlugin {
 	public boolean loginCheck(Player player){
 		// Check when user last logged in, only update username and group if over x hours ago
 		// TODO
+		permissions.hasGroupSupport();
 		return true;
 	}
 }
