+++ nameless-spigot/src/main/java/com/namelessmc/namelessplugin/spigot/commands/RegisterCommand.java
