+++ nameless-spigot/src/main/java/com/namelessmc/namelessplugin/spigot/player/PlayerEventListener.java
