+++ nameless-spigot/src/main/resources/config.yml
@@ -1,0 +1,12 @@
+# Nameless Plugin configuration
+
+# API URL
+# This needs to be the full API URL, including API key
+# For example http://yoursite.com/api/v1/API_KEY
+api-url: ""
+
+# Use reports?
+# If true, a /register command will be added to report users ingame
+# which will be added to the website's report system
+# Valid values = "true", "false"
+enable-reports: "true"
