+++ nameless-spigot/src/main/resources/plugin.yml
